+++ library/include/uniform_distribution_noise.hpp
@@ -6,7 +6,7 @@
 namespace cps {
     class UniformDistributionNoise : public Signal {
     public:
-        UniformDistributionNoise(int amplitude, int initialTimeSec, int durationSec);
+        UniformDistributionNoise(double amplitude, int initialTimeSec, int durationSec);
 
     public:
         double value(double /* x */) const override;
